rabbit fixes

+++ zea/data/convert/camus.py
@@ -159,10 +159,9 @@ def process_camus(source_path, output_path, overwrite=False):
         image_seq_polar.append(transform_sc_image_to_polar(image))
     image_seq_polar = np.stack(image_seq_polar, axis=0)
 
-    # Change range to [-60, 0] dB
+    # Change range to [-60, 0] dB — keep as float32, not uint8
     image_seq = translate(image_seq, (0, 255), (-60, 0))
     image_seq_polar = translate(image_seq_polar, (0, 255), (-60, 0))
-    image_seq_polar = image_seq_polar.astype(np.uint8)
 
     # Add y dimension (elevation) — CAMUS is 2D, so y=1
     image_seq_polar = np.expand_dims(image_seq_polar, axis=-1)
